extract test/utils into seperate functions, change test-runner to tape
