+++ memory.js
@@ -1,10 +1,11 @@
 var pull = require('pull-stream')
 var Filter = require('./filter')
 
-module.exports =  function createMemoryIndex (log, name) {
+module.exports = function createMemoryIndex (log, name) {
   function fullScan (log, opts) {
     return log.stream({
-      values: true, seqs: false,
+      values: true,
+      seqs: false,
       //TODO test coverage for live/old - the tests arn't right for live when the log starts as empty
       old: (opts.old !== false),
       live: (opts.live === true || opts.old === false),
@@ -13,7 +14,7 @@ module.exports =  function createMemoryIndex (log, name) {
   }
 
   console.error('flumeview-query:', name, 'in memory log or no indexes defined, will always use full scan, queries will likely be slow')
- return {
+  return {
     since: log.since,
     get: log.get,
     methods: { get: 'async', read: 'source' },
